+++ src/plugins/AB-Why3/preludes/cram.t/run.t
@@ -1,3 +1,0 @@
-  $ alt-ergo ../b-set-theory-prelude-2018-09-28.ae
-  $ alt-ergo ../b-set-theory-prelude-2020-02-28.ae
-